+++ y60/products/ViewACM/ScreenRegion.h
@@ -30,20 +30,20 @@ namespace video {
             virtual ~ScreenRegion(); 
 
             // access
-            void set(const EDGE& theEdge, const double & s, const double & t);
-            void get(const EDGE& theEdge, double & s, double & t) const;
+            void set(const EDGE& theEdge, const float & s, const float & t);
+            void get(const EDGE& theEdge, float & s, float & t) const;
             std::vector<float>  get(const EDGE& theEdge) const;
 
             // transform
             void rotate(const double & angle);
-            void stretch(const double & myXStretch, const double & myYStretch);
+            void stretch(const float & myXStretch, const float & myYStretch);
             void dump();
             void reset();
         private:
-            double _myTopLeft[2];
-            double _myTopRight[2];
-            double _myBottomRight[2];
-            double _myBottomLeft[2];
+            float _myTopLeft[2];
+            float _myTopRight[2];
+            float _myBottomRight[2];
+            float _myBottomLeft[2];
     };
     
 }
